wsd: allow longer timeout for child forking
We need to give the child more time than we
poll, so we would get a chance to register
prespawned children before we account for
any missing ones. Also, allowing only 5
seconds risks spawning too many children
on a slow, or somewhat loaded system.

This raises the child timeout to twice
the poll timeout, which is 5 seconds,
allowing for 10 seconds for spawning.

Reviewed-on: https://gerrit.libreoffice.org/80322
Tested-by: Andras Timar <andras.timar@collabora.com>

+++ wsd/LOOLWSD.cpp
@@ -3190,7 +3190,7 @@ int LOOLWSD::innerMain()
         }
         else
         {
-            const long timeoutMs = CHILD_TIMEOUT_MS * (LOOLWSD::NoCapsForKit ? 150 : 50);
+            const int timeoutMs = CHILD_TIMEOUT_MS * (LOOLWSD::NoCapsForKit ? 150 : 50);
             const auto timeout = std::chrono::milliseconds(timeoutMs);
             LOG_TRC("Waiting for a new child for a max of " << timeoutMs << " ms.");
             if (!NewChildrenCV.wait_for(lock, timeout, []() { return !NewChildren.empty(); }))
